Basic Searching and Default Sorting by Name
Ingredients and recipes are now default sorted by name. This was done by
changing the underlying data structure from being a HashSet to being a
ConcurrentSkipListSet (a skip list always maintains ordering). This comes
at the consequence that now insertion and deletion occur in O(log(n)). A
special thing to note here though is that the move to this structure
should make all data thread safe.

Furthermore, a rudimentary searching ability was added. This works by
filtering down the contents of the current view. Shown at the top will be
all elements that begin with the current search phrase and below will be
all elements that contain the phrase.

Other Minor Changes:
- Reformatted and cleaned up the async tasks. This was required as part
  of the move to ConcurrentSkipListSet.
- Abstract bases classes were created for the data managers -
  BaseDataManager - and the models - BaseModel. The addition of
  BaseDataManager was to reduce code commonality between RecipeData and
  IngredientData. The creation of BaseModel (which only requires a name)
  was to get BaseDataManager working properly.
- RecipeAdapter was modified to not store its data locally but instead to
  refer directly to the displayData of its specified data manager.