Fixed Recipes for the Current Ingredients
Recipes were not being displayed for the current ingredient. The exact
reason why is still unclear but it appears to be linked to either the
use of an `IN` statement for ingredient ID comparison against the
provided collection of ingredient IDs or the form of the collection of
ingredient IDs. Enough debugging time was put into this without success
so a new implementation was designed.

`getSuitableRecipes` no longer accepts a collection of ingredient IDs.
Instead, the SQLite query determines all ingredients that the user has
directly and then determines the collection of recipes where the user
has 'enough' of the ingredients for the recipe to be relavent.

Minor changes:
- `RecipesViewModel` no longer maintains a hashset of the available
  ingredient IDs.
- Relevancy for recipes was extended to also ensure that the number of
  matching ingredients is at least half of the total for the recipe.
- Suitable recipe ordering was extended to two stages. The first is
  still the number of missing ingredients. The new second stage is to
  order by number of matching ingredients. This is to encourage recipes
  that are more complex or larger to be shown to the user before more
  simple two ingredient snack or drink type recipes.